Add tests for deleteBudget

+++ pkg/handler/budgets.go
@@ -78,16 +78,27 @@ func (h *BudgetHandler) CreateBudget(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *BudgetHandler) DeleteBudget(w http.ResponseWriter, r *http.Request) {
-	id, _ := strconv.Atoi(r.PathValue("id"))
 	encoder := json.NewEncoder(w)
+	parts := strings.Split(r.URL.Path, "/")
+	id, err := strconv.Atoi(parts[len(parts)-1])
+
+	if err != nil {
+		utils.ErrRes(w, errors.Generic400Err, http.StatusBadRequest)
+		return
+	}
+
+	if _, err := h.repo.GetBudget(id); err == errors.Budget404Err {
+		utils.ErrRes(w, errors.Budget404Err, http.StatusNotFound)
+		return
+	}
 
 	if err := h.repo.DeleteBudget(id); err != nil {
-		utils.ErrRes(w, err, 500)
+		utils.ErrRes(w, err, 400)
 		return
 	}
 
 	payload := models.GenericPayload{Msg: "Budget deleted"}
-	w.WriteHeader(http.StatusAccepted)
+	w.WriteHeader(http.StatusNoContent)
 	encoder.Encode(payload)
 }
